Return proper exit code for make check
In the test-go.sh the last command is popd which will be the exit
code of the script. This means that even if running the actual tests
will fail, the exit code of the script will always be 0.

This change will also fix travis ci to fail when tests are failing.

Signed-off-by: Alin-Gheorghe Balutoiu <alinbalutoiu@gmail.com>