Support multiple networks - minion only
The purpose of this commit is to extend ovn-kubernetes to support multiple networks.
The secondary networks will be added using multus (or any other solution complies with the
de-facto standard).

It means the pods annotation should contain the required network attachments, MACs and IPs
under -'k8s.v1.cni.cncf.io/networks'.

```apiVersion: v1
kind: Pod
metadata:
  name: pod-with-multiple-ovn-networks
  annotations:
    k8s.v1.cni.cncf.io/networks: '[
            {
             "name": "blue-network-attachment",
             "mac": "0a:00:00:00:00:70",
             "ips": "192.168.0.2/24"
            },
            {
             "name": "red-network-attachment",
             "mac": "0a:00:00:00:00:71",
             "ips": "10.0.0.2/24"
            }
    ]'
...
```

The network attachment CRD should contain the name of the logical switch represents the
logical network in the config section.

```apiVersion: "k8s.cni.cncf.io/v1"
kind: NetworkAttachmentDefinition
metadata:
  name: “blue-network-attachment”
spec:
  config: '{
    “name”: “blue-network”
    "cniVersion": "0.3.1",
    "type": "ovn-k8s-cni-overlay",
    "Ipam":{},
    "dns":{}
  }'
```

Once the pod will be started, multus will invoke on the minion the binary 'ovn-k8s-cni-overlay'
per each pod's network attachment with the type 'ovn-k8s-cni-overlay'.

The ovn-kubernetes minion code is changed in the following way -

1. **Multus invokes the 'ovn-k8s-cni-overlay' per each interface** (of course only network
interfaces that are using network attachment with the type - 'ovn-k8s-cni-overlay').
2. **The CNI binary ('ovn-k8s-cni-overlay') reads the name of the network/switch from the
cni.config and the MAC and the IP from the cni.env["CNI_ARGS"].
3. A veth pair is created in the container and then the host end is moved into host netns.
4. **A new port is created on “br-int”** connecting the host-end to the switch **with "
external_ids:iface-id: namespace_podName_logicalSwitchName”** (the logicalSwitchName is
taken from the NetworkAttachment and passed by multus to the minion binary (same as the
other values from the config section of the network attachment).

Notes:
1. Currently multus has two bugs/unimplemented parts -
 - The "ips" addresses specified in the pod network annotation are not passed to the CNI binary.
 - Multus expects the "ips" to be a string although according to the de-facto standard
documentation, it should be an array of strings.
 I will send multus a PR to fix the issues.
2. According to the documentation of the de-facto standard, the subnet is not specified as part
of the "ips".
I need to discuss about it with the maintainers of the standard, but currently in ovn-kubernetes,
I assume the format of each ip is ip/prefix.
3. Although the pod-network annotation is called ips, only one IP is supported (it is passed to
the cni binary on cniRequest["ENV"]["CNI_ARGS"]["IP"]).
4. An external system will be responsible for adding the logical switches/ports/... to the ovn
DB. The name of the ports should be "namespace_podName_logicalSwitchName"
5. The secondary ovn-kubernetes networks can be added whether ovn-kubernetes is the
 primary network or not.
6. Interfaces on the same pod connected to the same network/logical switch are currently not
supported (since the name of logical port is namespace_podName_switchName and we cannot
have multiple ports with the same name).