windows: Fixes MAC fetching during gateway-init
There is a known issue with OVS not correctly picking up the
physical network interface MAC address.

Gateway-init will fail on the Windows node when trying to add
the router port with a different MAC than the initial one. This
happens after node reboot.

As a workaround, fetch the MAC address via powershell if
it is not picked up from OVS.

Signed-off-by: Alin-Gheorghe Balutoiu <alinbalutoiu@gmail.com>
Signed-off-by: Alin Gabriel Serdean <aserdean@ovn.org>