remove init-gateway command from ovn-kube-util(1) CLI
OVN L3 Gateway functionality is now enabled from ovnkube binary. We are not using this binary
anymore. Also, this program is out-of-sync with the gateway features (br-localnet, and others)
that we have added in the past few months. It is better to remove this code then to keep it
here and let it stale.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>