RunOVNNbctlHA() makes an incorrect assumption on which scheme to use
RunOVNNbctlHA incorrectly assumes that if a single server is provided
for ovn-nb, then OVN NB is available locally and it switches the
connection to use AF_UNIX. It could be that the OVN DBs are external
to K8s cluster and may not be available locally on the master node
(or that the OVN DB is configured as active-standby or active-backup
nodes outside of K8s cluster)

If users are deploying a single instance of master and OVN DBs and
if they really need to use UNIX scheme for more reliability, then they
can always start the ovnkube.sh with --nb-address and --sb-address
set to nothing. This will automatically falls back to UNIX scheme.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>