Gateway: do not require the interface name to be in bridge name
Currently, the ovnkube in node mode requires that the OVS bridge name
start with br<gw_inteface_name> for the shared gateway interface option.
This is hard to enforce if the bridge is already created in the host by
the user.

The following changeset adds a new key called 'bridge-uplink' to the
external_ids column of the bridge table. This key captures the shared
gateway interface. With that, if the user creates the bridge ahead of
time, then they will need to populate the bridge-uplink to the shared
gateway interface value and they can name the bridge whatever they want
(br-eth2, ovs-br, admin-br, and so on).

Also, when ovnkube creates a bridge using util.NICtoBridge() code it
will also populate the bridge-uplink key.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>