gateway: add access to tagged external network
All the external logical switches that connect L3 gateway to external
network assumes that the underlay is untagged. This may not be the case.
The underlay network could be tagged and therefore we need a way to
specify a VLAN ID.

This commit adds a new argument to ovnkube called --gateway-vlanid. The
default value will be 0, and therefore untagged. If the external network
is tagged, then the packets leaving the POD must be tagged as well and
one should populate this option with the right VLAN tag.

This option is valid only for Shared Gateway Interface and Spare
Gateway Interface.

Fixes #583

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>