ovn: wait until node switch appears before starting pods on that node
Wait a while for the node switch to show up before failing to create
a pod scheduled on that node. This can happen due to a race between
the pod & node events when the scheduler assigns a pod to the node
before we've had a chance to process the node addition.

Signed-off-by: Dan Williams <dcbw@redhat.com>