[Ansible][contrib] Add workaround for lineinfile concurrency issues
Ansible will try to manage all of the machines
referenced in a play in parallel by default.

The module lineinfile has conccurency issues
when multiple hosts are delegating to the
same host and the same file. This issue
is tracked upstream in ansible:
https://github.com/ansible/ansible/issues/30413

As a temporary workaround, the hosts file will
be updated by all the nodes one by one instead
of doing it in parallel.

Signed-off-by: Alin Balutoiu <alinbalutoiu@gmail.com>