framework for providing high availability for OVN DBs
The idea here is to have 3 sets of PODs:

1. ovnkube-db POD -- runs both OVN NB DB and OVN SB DB. Other PODs will wait
until this POD comes up since it has DBs.  This will be facilitated using
headless service without label selectors -- named ovnkube-db. The endpoint for
this service will be populated by ovnkube.sh once both the DBs are up.

2. ovnkube-master POD -- runs ovn-northd and ovnkube in master mode. This POD
depends on ovnkube-db and becomes aware of the availability of the DB POD by
checking the ovnkube-db Endpoint object (just like how we do it today).

3. ovnkube-node POD – no changes and will have containers like today.

Fixes #648

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>
Signed-off-by: Yun Zhou <yunz@nvidia.com>