config: server side configuration must be handled outside of ovnkube
The OVN central DB's configuration such as setting it up for SSL
connections or TCP connections must be done outside of the ovnkube.
It is not possible to do it in ovnkube because not all the deployments
will run ovnkube and OVN DBs on the same host, so one cannot assume the
availability of UNIX socket path.

Fixes #644

In the case of daemonsets, the entrypoint for the containers (aka
ovnkube.sh) will handle it. In the case of Vagrant, the k8s-master
script will handle it.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>