wait on logical switch creation before creating the management port
currently, ovnkube-node checks for the presence of "ovn_host_subnet"
node annotation, in a loop, to ensure that ovnkube-master has created
node's logical_switch and the corresponding subnet.

This causes an issue whenever we delete ovn-k8s CNI components and
recreate them. This is because the node annotation is not removed as
the node itself has not been deleted from the K8s cluster. The issue is
that ovnkube-node sees the ovn_host_subnet and tries to create the
management port, however the logical switch would not have been created
it. So, it is better to depend on the OVN DB for the presence of OVN
logical switch instead of the annotation.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>