iptables: issues with usage of append() and iptable rules ordering
Per append() documentation if the slice has sufficient capacity, then it
would reslice the 1st argument to accommodate the new elements. This
was causing a logic error in how the iptable rules were getting inserted
at position 1 in a chain.

Also, the rule ordering was incorrect if the chain already had one
rule in it.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>