Add unidling support.
When all pods for a service VIP are removed, this used to result in the
load balancer for the VIP being removed from OVN.

With this change, we add an option to instead leave the VIP configured,
but have the backends in the OVN load balancer empty.

We also enable the controller_event northbound option. Now, with empty
load balancer backends, if OVN receives traffic for a load balancer VIP,
it will raise an empty_lb_backends event in the southbound database.

This patch adds an EventChecker method for the Controller that polls the
southbound database every 5 seconds to find if there are any unhandled
empty_lb_backend events. If there are any, then the event is parsed and
we send a NeedPods event.

A CMS can respond to this NeedPods event by spinning up new pods. This
way, pods can be created to handle the traffic being sent to the load
balancer.

Signed-off-by: Mark Michelson <mmichels@redhat.com>
Signed-off-by: Lorenzo Bianconi <lorenzo.bianconi@redhat.com>