OVN: Stop putting the token on the cmdline
Modify ovnkube to use environment variables and service account files to
provide Kubernetes configuration in addition to the command line
options.

The value is set from the command line option, if present, or the
environment variable, if present, or the service account file, where
available.

--k8s-apiserver  K8S_APISERVER
--k8s-kubeconfig KUBECONFIG
--k8s-cacert K8S_CACERT /var/run/secrets/kubernetes.io/serviceaccount/ca.crt
--k8s-token K8S_TOKEN /var/run/secrets/kubernetes.io/serviceaccount/token

The ovnkube.sh script is changed to no longer pass ovnkube --k8s-apiserver
--k8s-cacert --k8s-token since they are availabe as described above.

Jira: SDN-304
https://jira.coreos.com/browse/SDN-304

Signed-off-by: Phil Cameron <pcameron@redhat.com>