node processes should get their connections via a Service
Rather than distributing the database urls via ConfigMaps, they should be obtained via a Service/Endpoint.
As endpoints change, the ovn-kube process can update ovn and openvswitch accordingly

includes a small style update. Golang style says that all the letters in an acronyms should have consistant case.
Since I used this function in my PR I updated corrected the style for UpdateOvnNodeAuth -> UpdateOVNNodeAuth

Added unit tests for validateOVNConfigEndpoint() and watchConfigEndpoints() to ensure that the northbound and southbound addresses are configured with changes to the endpoint and that the endpoint has the correct information

Signed-off-by: Jacob Tanenbaum <jtanenba@redhat.com>