Change the value for selecting the master node
Jira: ENTNET-1377

For most current Kubernetes master node, the label value of
"node-role.kubernetes.io/master" is just ""(empty), not "true".

So when using the nodeSelector to choose the master of K8s cluster,
we should use just an empty value "" instead of a "true" value, which
means if there is a label "node-role.kubernetes.io/master" set in the node,
it's the desired master node.

If the value of "true" is used, there may be NO approriate node
that can be found to run the pod when deploying the pod.

Signed-off-by: trevor tao <trevor.tao@arm.com>