remove the experimental --ha option from ovn-k8s
--ha provided a way to rebuild OVN DB in the case when the DBs got
corrupted or got lost. However, we now have a way to provide the OVN DB
HA either natively through K8s or outside of k8s. As a result, --ha
option as a OVN DB HA is not that useful. Also, that part of the code
is unmaintained and has not been tested in a while.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>