cluster: Do health checking for LoadBalancer services
LoadBalancer services can specify a health check port on which the
proxy implementation is expected to run a health check service
indicating whether the service is available on that node. (This is
used by the cloud load balancer to know when it is OK to route to a
particular node.)

pkg/kube/healthcheck/ is a copy of kubernetes's
pkg/proxy/healthcheck/, with a few dependency-related simplifications.

Signed-off-by: Dan Winship <danw@redhat.com>