OVN: use config file via ConfigMap rather than environment variables
The ovn-config configMap can be mounted into the container and
values taken from there rather than using the environment variables.

The config-file for ovnkube can be passed by a configMap and mounted
into the container. This limits the need for environment variables.

This change in API results in

OVN_DAEMONSET_VERSION 4
Version 3 is still supported.
Eliminated v1 and v2 support.

The environment variables in V3 are preserved for existing users.

See also:
https://github.com/openshift/cluster-network-operator/pull/217

SDN-456
https://jira.coreos.com/browse/SDN-456

Signed-off-by: Phil Cameron <pcameron@redhat.com>