OVN NB/SB DB HA using Corosync/Pacemaker
Start OVN DBs in Active/Standby HA mode using Corosync and Pacemaker.
Both the Northbound/Southbound DBs run in a single container in the
ovnkube-db pod. This is done so that it simplifies setting up
corosync/pacemaker as we just need one cluster to run both the DBs.

The OVN NB/SB DB will be available behind a Virtual IP, and in the
simple case the value of the Virtual IP could be from the K8s Node
subnet. Virtual IP address from a routable subnet could also be used.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>