Add HA support for ovnkube masters.
With this patch, we can run multiple ovnkube in master mode.
Only one will be active. The client-go/tools/leaderelection module
is used to elect a leader.

It is expected that the HA of OVN ovsdb-servers if desired is managed separately.

When the ovnkube-master is started, it participates in the
election. The ovnkube-master which wins the election starts monitoring the
pod events. The followers do nothing.

Signed-off-by: Numan Siddique <numans@ovn.org>