Add use of daemon mode for ovn-nbctl.
Running ovn-nbctl typically results in a new connection being made to
the northbound database each time. This new connection results in
ovn-nbctl downloading the entire contents of the northbound database,
which becomes inefficient as the network grows large.

With this change, we instead run an ovn-nbctl daemon. The daemon
maintains a constant connection to the northbound database and receives
updates as the network changes. We then use ovs-appctl to connect to the
ovn-nbctl daemon whenever we normally would have run ovn-nbctl. This
results in a much more performant means of updating the network.

Only the ovn-kubernetes master connects uses the daemon. The
ovn-kubernetes-node uses ovn-nbctl in the traditional sense.