gateway/local: gateway-originated packets must return via the gateway
When a pod accesses a Service and the selected endpoint happens to be
on the local machine, the packet exiting the gateway interface (br-nexthop)
will be SNAT-ed to the host's IP. The return packet will be un-NATed
back to the originating Pod's IP. But since that IP is in this host's
hostsubnet, the kernel routes that packet to the management port and
stuff doesn't work.

Fix that by marking all connections originating from br-nexthop, and
sending reply packets to a new routing table that sends them along
to br-nexthop rather than the management port.

Signed-off-by: Dan Williams <dcbw@redhat.com>