master: don't block waiting for gateway load balancer creation
When a node was added, the master-side NodePort code would block for
up to 5 minutes waiting for the gateway load balancers to be created
(which would happen only after the "ovnkube --node" process was up and
running on the new node, which would happen only after the
ovn-kubernetes image had been pulled, etc).

Since all event handlers of a given type are run serialized in
ovn-kubernetes, blocking in the NodePort Node-added handler would also
prevent the primary Node-added handler from running, preventing it
from registering any other new nodes.

The net result of this is that if you created multiple new nodes at
once, and deployed pods to them which could tolerate NotReady, then on
the second and later nodes, those pods would never successfully come
up, because the Pod-added handler in the master would time out before
the Node had been fully registered.

To fix this, we change the NodePort setup code to just return
immediately if the load balancers don't exist yet, and try again the
next time we get an UPDATE event for the Node (taking advantage of the
fact that Nodes' Status fields are updated quite frequently).

https://bugzilla.redhat.com/show_bug.cgi?id=1729091

Signed-off-by: Dan Winship <danw@redhat.com>