ovnkube.sh: optionally run as non-root user
When the shell script sees the OVS_USER_ID parameter, it will run the
various daemons as that user.  The user-info is specified in the standard
user:group notation.

As an implementation detail, using libcap-ng to do the domain transition for
the ovn containers seemed to cause issues, possibly with the capability
bits being retained.  For now, use setpriv as a workaround, but it could be
addressed in the future without impacting users.

Signed-off-by: Aaron Conole <aconole@redhat.com>