remove other-config:exclude_ips key once that IP address is allocated
ovn-northd complains with a warning message "Duplicate IP set" when the
IP address set in logical switch's other_config:exclude_ips key is
statically allocated to a logical switch port. To eliminate that
warnings we'd need to remove the exclude_ips configuration in the same
transaction as the creation of the management port.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>