ovnkube-master and ovnkube-db pods shouldn't have hostPID set to true
Currently, those two pods have hostPID set to true. So, this means that
they can see all the processes in host process namespace. This is not
good from the security perspective.

Commit 59c345931f9d (use ovs-appctl to check for process readiness and
health) added support to check process readiness and health using
a control socket instead of a PID.

ovnkube-node pod cannot be yet moved to hostPID:false since we open
network namespace (PID in host process namespace) from within the pod
while adding an interface to the pod. so, this pod will be tackled in
a separate commit.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>