ovn-nbctl daemon mode until we get golang OVSDB bindings
Every invocation of ovn-nbctl command connects to northbound database
running on a remote server, retrieves a partial copy of the database
that is complete enough to do its work, sends a transaction request to
the server, and receives and processes the server’s reply. In common
interactive use, this is fine, but if the database is large, then the
command takes a long time to complete and yields poor performance.

The solution to this problem is to use golang OVSDB bindings. However,
In the absence of the bindings the next best solution is to use
ovn-ncbtl in the daemon mode.

In a "daemon mode", the user first starts ovn-nbctl running in the
background and afterward uses the daemon to execute operations. The
client needs to use the control socket and set the path to the control
socket in environment variable OVN_NB_DAEMON.

Changes have been made such that it can be used in non-daemonset mode as
well.

Also, this will be used for ovn-nbctl invocations in ovnkube --master
mode only.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>