remove TEMP HACK - WORKAROUND verbiage from ovnkube.sh
Running "local" gateway mode is not a HACK. It is one of the ways to
achieve North-South traffic (aka traffic external to logical network).

If operators want to run "shared" gateway mode, then it is well
understood that they shouldn't be running ovs-server container in the
ovnkube-node POD. The expectation is that the OVS bridge with
right physical NIC as one of the bridge's port is already created in
the K8s node and that ovs-vswitchd and ovsdb-server processes are
running in the node.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>