enable port security on the pod's logical switch port
The port security is currently disabled. This means that the packets
with any ip/mac can be received/sent by the pod. A pod can do both MAC
and IP spoofing. However, by enabling port security we can limit the
addresses from which a logical port may send packets and to which it
may receive packets. This can be enabled by setting the
logical_switch_port's port_security column to the port's (MAC, IP)
address.

We cannot enable port security on K8s management logical switch port
since packets meant for local k8s node are forwarded through that port.
So, it receives packets with destination IP address that does not
belong to it.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>