Add support for MAC attribute in CNI_ARGS
The MAC attribute can be can be passed via CNI_ARGS into the
ovn-kubernetes CNI plugin. If it's passed, then the plugin is supposed
to enforce it and that is not the case with OVN CNI plugin. This commit
fixes the issue.

The multus-CNI reads the custom MAC off of the annotation on the POD
spec defined like this below:

annotations: v1.multus-cni.io/default-network: | [
    {
        "namespace": "kube-system",
        "name": "ovnkube",
        "mac": "aa:bb:cc:dd:ee:ff"

    }
]

The OVN interface that gets added to the POD will have that MAC instead
of OVN chosen MAC.

Fixes #776

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>