update the default route when the next hop node is deleted
We add a default route to ovn_cluster_router that points to a L3
gateway router as the next hop. When the node that hosts the L3 gateway
router is deleted from the cluster, the default route still points to
the same l3 gateway router chassis. We also need to move all the
External IPs related Load Balancer rules to a different default gateway
router.

Signed-off-by: Zhen Wang <zhewang@nvidia.com>