do not copy over loopback CNI plugin from the onvkube-node container
all of the docker images for ovnkube-node container doesn't have
loopback CNI plugin installed, so there is no point in trying to copy
over this binary as it will result in an error.

furthermore, the `loopback` CNI plugin will need to be installed by
the deployer on each of the kubernetes node as part of the node bring
up. this is one of the requirements of kubernetes. (See Kubernetes
CNI network plugin requirements on kubernetes.io)

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>