dist: Remove k8s_apiserver parameter for containerized ovnkube
When ovnkube is running in containers inside the Kubernetes cluster, it
doesn't have to be aware of the kubeapi-server address. Kubernetes
client should work with the in-cluster config and kubectl inside pods
should work out of the box.

Hardcoding the kubeapi-server IP might be problematic if master nodes
get changed, redeployed or if the network configuration of Kubernetes
nodes gets changed.

Signed-off-by: Michal Rostecki <mrostecki@opensuse.org>