exit early from the entrypoint script on error
The process_ready() function read's the process pid from the pidfile and
ensures that the pid is running. If the pidfile is not present, then it
keeps checking (through wait_for_event code) for the pidfile to show up.

However, there is a possibility that the process crashed and no pidfile
was written. Now, the above logic will keep trying for 400s and give up.
The status of the container will be Running but the process has died
within the first few seconds.

Signed-off-by: Yun Zhou <yunz@nvidia.com>
Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>