Add OVN-KUBE-SNAT-MGMTPORT chain in iptables nat table
Packets entering the OVN pipeline from the host or hostNetwork pods are
now SNAT'ed to management port IP. This is done so that the response
packets come back through the management port IP itself. However, the
rule is now just added to POSTROUTING chain. This fix moves the SNAT
rule to its own chain called OVN-KUBE-SNAT-MGMTPORT and delete the
chain on node deletion.

Signed-off-by: Zhen Wang <zhewang@nvidia.com>