remove support for spare interface gateway mode
The spare interface gateway mode allows one to take a spare interface
on the node and attach the interface directly to br-int. When you do
that, all the North-South traffic from the Pod will exit out through
the spare interface.

However, the issue with this mode is that the NodePort services aren't
supported (and therefore LoadBalancer type services).

This is a severe drawback with this mode, and we know that no one in
the community is using this mode. So, we should just remove this mode
to facilitate some of the upcoming PRs (annotate node with L3 gateway
information and move gateway creations to master node).

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>