move master related receivers in OvnClusterController to Controller
Right now the master operations are under two different structures
and its receivers. This commit combines all the master operations
under the Controller struct in `ovn` package.

By doing so, all the node operations are under the OvnClusterController
in the `cluster` package.

So, there will be two controllers -- master (runs in ovnkube-master
process) and node (runs in ovnkube-node process).

In the next PR, I will rename `Controller` struct to OvnMasterController
and `OvnClusterController` struct to OvnNodeController.

Signed-off-by: Girish Moodalbail <gmoodalbail@nvidia.com>